fix(build): add .js extensions to ESM imports in lib/metrics.ts and related Node files
