+++ lib/metrics-shared.js
@@ -1,21 +1,4 @@
-export type Direction = 'up_is_better' | 'down_is_better'
-
-export type YearValue = { year: number; value: number }
-export type AnyPoint = { year?: number; date?: number; value?: number; coverage?: number; [k: string]: any }
-
-export type Metric = {
-  id: string
-  name: string
-  domain: string
-  unit: string
-  direction: Direction
-  source: string
-  dataPath?: string
-  subtitle?: string
-  detailPath?: string
-}
-
-export const METRICS: Metric[] = [
+export const METRICS = [
   { id: 'co2_ppm', name: 'CO₂ concentration', domain: 'Climate & Environment', unit: 'ppm', direction: 'down_is_better', source: 'NOAA/ESRL' },
   { id: 'life_expectancy', name: 'Life expectancy', domain: 'Health & Wellbeing', unit: 'years', direction: 'up_is_better', source: 'WHO/World Bank' },
   { id: 'internet_use', name: 'Individuals using the internet', domain: 'Education & Digital', unit: '%', direction: 'up_is_better', source: 'ITU' },
@@ -85,7 +68,7 @@ export const METRICS: Metric[] = [
 
 // Adapter to align validator-friendly coverage feeds with UI expectations.
 // Keep schema expectations in sync with scripts/validate-datasets.cjs.
-export function toYearValuePercent(raw: AnyPoint[]): YearValue[] {
+export function toYearValuePercent(raw) {
   if (!Array.isArray(raw)) return []
   const out = raw
     .map(d => {
